+++ scripts/iemre/init_prism_dailyc.py
@@ -38,7 +38,7 @@ def init_year(ts):
     nc.createDimension("nv", 2)
 
     # Setup Coordinate Variables
-    lat = nc.createVariable("lat", np.float, ("lat",))
+    lat = nc.createVariable("lat", float, ("lat",))
     lat.units = "degrees_north"
     lat.long_name = "Latitude"
     lat.standard_name = "latitude"
@@ -47,11 +47,11 @@ def init_year(ts):
     # Grid centers
     lat[:] = prism.YAXIS
 
-    lat_bnds = nc.createVariable("lat_bnds", np.float, ("lat", "nv"))
+    lat_bnds = nc.createVariable("lat_bnds", float, ("lat", "nv"))
     lat_bnds[:, 0] = prism.YAXIS
     lat_bnds[:, 1] = prism.YAXIS + 0.04
 
-    lon = nc.createVariable("lon", np.float, ("lon",))
+    lon = nc.createVariable("lon", float, ("lon",))
     lon.units = "degrees_east"
     lon.long_name = "Longitude"
     lon.standard_name = "longitude"
@@ -59,11 +59,11 @@ def init_year(ts):
     lon.axis = "X"
     lon[:] = prism.XAXIS
 
-    lon_bnds = nc.createVariable("lon_bnds", np.float, ("lon", "nv"))
+    lon_bnds = nc.createVariable("lon_bnds", float, ("lon", "nv"))
     lon_bnds[:, 0] = prism.XAXIS
     lon_bnds[:, 1] = prism.XAXIS + 0.04
 
-    tm = nc.createVariable("time", np.float, ("time",))
+    tm = nc.createVariable("time", float, ("time",))
     tm.units = "Days since %s-01-01 00:00:0.0" % (ts.year,)
     tm.long_name = "Time"
     tm.standard_name = "time"
@@ -72,7 +72,7 @@ def init_year(ts):
     tm[:] = np.arange(0, int(days))
 
     p01d = nc.createVariable(
-        "ppt", np.float, ("time", "lat", "lon"), fill_value=1.0e20
+        "ppt", float, ("time", "lat", "lon"), fill_value=1.0e20
     )
     p01d.units = "mm"
     p01d.long_name = "Precipitation"
@@ -81,7 +81,7 @@ def init_year(ts):
     p01d.description = "Precipitation accumulation for the day"
 
     high = nc.createVariable(
-        "tmax", np.float, ("time", "lat", "lon"), fill_value=-9999.0
+        "tmax", float, ("time", "lat", "lon"), fill_value=-9999.0
     )
     high.units = "C"
     high.long_name = "2m Air Temperature Daily High"
@@ -89,7 +89,7 @@ def init_year(ts):
     high.coordinates = "lon lat"
 
     low = nc.createVariable(
-        "tmin", np.float, ("time", "lat", "lon"), fill_value=-9999.0
+        "tmin", float, ("time", "lat", "lon"), fill_value=-9999.0
     )
     low.units = "C"
     low.long_name = "2m Air Temperature Daily High"
